+++ packages/runtime/src/helpers/redirects.ts
@@ -7,7 +7,7 @@ import type { PrerenderManifest, SsgRoute } from 'next/dist/build'
 import { outdent } from 'outdent'
 import { join } from 'pathe'
 
-import { HANDLER_FUNCTION_PATH, HIDDEN_PATHS, ODB_FUNCTION_PATH } from '../constants'
+import { HANDLER_FUNCTION_PATH, ODB_FUNCTION_PATH } from '../constants'
 
 import { isAppDirRoute, loadAppPathRoutesManifest } from './edge'
 import { getMiddleware } from './files'
@@ -26,14 +26,6 @@ import {
 
 const matchesMiddleware = (middleware: Array<string>, route: string): boolean =>
   middleware.some((middlewarePath) => route.startsWith(middlewarePath))
-
-const generateHiddenPathRedirects = ({ basePath }: Pick<NextConfig, 'basePath'>): NetlifyConfig['redirects'] =>
-  HIDDEN_PATHS.map((path) => ({
-    from: `${basePath}${path}`,
-    to: '/404.html',
-    status: 404,
-    force: true,
-  }))
 
 const generateLocaleRedirects = ({
   i18n,
@@ -280,8 +272,6 @@ export const generateRedirects = async ({
   const { dynamicRoutes, staticRoutes }: RoutesManifest = await readJSON(
     join(netlifyConfig.build.publish, 'routes-manifest.json'),
   )
-
-  netlifyConfig.redirects.push(...generateHiddenPathRedirects({ basePath }))
 
   if (i18n && i18n.localeDetection !== false) {
     netlifyConfig.redirects.push(...generateLocaleRedirects({ i18n, basePath, trailingSlash }))